chore: cleanup finder
